More test separation
